// src/app/core/services/snc/model/networkType.ts
/**
 * SNC
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export type NetworkType = 'Telegram' | 'VK';

export const NetworkType = {
    Telegram: 'Telegram' as NetworkType,
    VK: 'VK' as NetworkType
};